tsract_exp/requirements.txt
numpy==1.19.5
opencv-contrib-python==4.5.1.48
Pillow==8.1.0
pytesseract==0.3.7